+++ examples/customizations/marytts_mp3_generator/__init__.py
@@ -6,7 +6,7 @@ from faker_file.providers.base.mp3_generator import BaseMp3Generator
 from speak2mary import MaryTTS
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("MaryTtsMp3Generator",)