+++ src/faker_file/base.py
@@ -16,7 +16,7 @@ from faker.providers.python import Provider
 from .helpers import wrap_text
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = (
     "BytesValue",