+++ src/faker_file/contrib/image/pil_snippets.py
@@ -1,5 +1,5 @@
 from ..pdf_file.pil_snippets import *  # noqa
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"