+++ src/faker_file/providers/augment_file_from_dir/extractors/tika_extractor.py
@@ -7,7 +7,7 @@ from tika import parser
 from ...base.text_extractor import BaseTextExtractor
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("TikaTextExtractor",)