+++ src/faker_file/providers/base/text_augmenter.py
@@ -1,5 +1,5 @@
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("BaseTextAugmenter",)