+++ src/faker_file/providers/base/text_extractor.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Union
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("BaseTextExtractor",)