+++ src/faker_file/providers/mp3_file/generators/edge_tts_generator.py
@@ -7,7 +7,7 @@ import edge_tts
 from ...base.mp3_generator import BaseMp3Generator
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("EdgeTtsMp3Generator",)