+++ src/faker_file/providers/pdf_file/generators/reportlab_generator.py
@@ -17,7 +17,7 @@ from ....constants import DEFAULT_FONT_NAME, DEFAULT_FONT_PATH
 from ...base.pdf_generator import BasePdfGenerator
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2022-2023 Artur Barseghyan"
+__copyright__ = "2022-2025 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = ("ReportlabPdfGenerator",)